refactor: move ScreenCover into its own first class directory
